requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1
django-haystack==2.8.1
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-password-reset==2.0
django-redis==4.11.0
django-simpleui==4.0.2
idna==2.9
importlib-metadata==1.6.1
Markdown==3.2.2
Pillow==7.1.2
Pygments==2.6.1
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.2
requests==2.23.0
urllib3==1.25.9
zipp==3.1.0
